Update verbs to be grammatically correct
This commit updates the verbs in the sentences for dirname, basename, and extname to be grammatically correct. The corrected sentences now read:

dirname: gets the parent folder of a file.
basename: gets the filename part.
extname: gets the file extension.
These changes improve the clarity and accuracy of the language used in the README file.

Signed-off-by: Tobechukwu David Uka <54605363+ukadavid@users.noreply.github.com>